Remove unused factory method
The factory method contains references to game modules names as
Games_Chess_$type and Games/Chess/$type, which are no longer part
of the repository (prior commits by other developers removed them).

It is therefore deemed safe to remove references to those modules and
the factory class that was loading them.